move to dedicated dynamic loaded component
